Better sync command and updated faces
The new sync command does away with the output buffer. So hopefully no
more annoying questions about whether to open it in a separate buffer.

Also update the faces so that the undo-visualize-tree looks like the
magit diffs. As part of this upgraded my wombat theme (moved some
configurations from the init.el into the theme). Still not sure how to
have the values of the colors be variables (like the macro used in
hc-zenburn-theme).

Lastly, company-dabbrev will no longer suggest completions for
numbers (it was doing it in org-mode).

Change-Id: I822f4189ec11422aaeb2e739cc90a6390169818f